requirements.txt
pyyaml
ollama
pydantic
pydantic-ai
pandas
matplotlib
gradio